Have decided to diverge from the tutorial I was following super early as I've come up with a method of sorting cards, storing and recalling their value, and managing decks that I've started implementing early.
I added an original C# script called card manager that contains utility functions that check if a card has been played and stores a list of all played cards. I also added the return values for my randomized face sprites to a giant switch statement in CardModel.

My plan is to have each card draw spawn a Card Object and attempt to assign a cardIndex value to it based on a Random.Range roll, utilizing my utility function to check whether or not the randomly generated index has been played, and if so rerolling. I figure that using a List of ints to store deck data and handling all the card values and operations through logic will keep my project a bit cleaner and make reshuffling the deck with cards still in play simple.

I may end up diverging from this plan if it is too resource intensive or complicated down the line, but as of right now I'm quite confident that it will work well.