scrapped almost everything from the previous version of this project. there are now 52 discrete card objects representing a full standard deck
most of the old scripts are gone and the cardModel script is different. I kept the cardFlipper but I have a different way of determining what face to show as I no longer am instancing card faces and values off of a single card object, but planning on spawning discrete card objects that are organized in a List called Deck, yet to be implemented.