cards "flip over" upon spawn now. This will be disabled for one dealer card, otherwise all cards will "flip over" when spawned.
next task is creating a "Deal" routine for the beginning of each turn cycle in which four cards are dealt, two to playerCards and two to a new List called dealerCards that will be implemented. after basic deal is working I plan to implement a cardValue check to keep track of what the player and dealer are showing. after that is the Hit/Stay cycle for the player and then basic dealer AI and dealer turn.